feat: add getCarts usecase

+++ usecase/cart_usecase.go
@@ -1,14 +1,17 @@
 package usecase
 
 import (
+	"git.garena.com/sea-labs-id/batch-06/ferza-reyaldi/stage01-project-backend/dto"
 	"git.garena.com/sea-labs-id/batch-06/ferza-reyaldi/stage01-project-backend/entity"
 	er "git.garena.com/sea-labs-id/batch-06/ferza-reyaldi/stage01-project-backend/error"
 	"git.garena.com/sea-labs-id/batch-06/ferza-reyaldi/stage01-project-backend/repository"
+	"gorm.io/gorm"
 )
 
 type CartUsecase interface {
 	AddToCart(userId int, courseId int) error
 	RemoveFromCart(userId int, courseId int) error
+	GetCarts(userId int) ([]*dto.CartsResponse, error)
 }
 
 type cartUsecaseImpl struct {
@@ -30,7 +33,6 @@ func NewCartUsecase(c *CartUConfig) CartUsecase {
 const CoursePrice = 150_000
 
 func(u *cartUsecaseImpl) AddToCart(userId int, courseId int) error {
-
 	_, err := u.courseRepository.FindById(courseId)
 	if err != nil {
 		return err
@@ -60,10 +62,13 @@ func(u *cartUsecaseImpl) AddToCart(userId int, courseId int) error {
 }
 
 func(u *cartUsecaseImpl) RemoveFromCart(userId int, courseId int) error {
-
 	cartedCourse, err := u.cartRepository.Find(userId, courseId)
 	if err != nil {
 		return err
+	}
+
+	if cartedCourse.Id == 0 {
+		return gorm.ErrRecordNotFound
 	}
 
 	err = u.cartRepository.Delete(cartedCourse.Id)
@@ -72,4 +77,23 @@ func(u *cartUsecaseImpl) RemoveFromCart(userId int, courseId int) error {
 	}
 
 	return nil
+}
+
+func(u *cartUsecaseImpl) GetCarts(userId int) (carts []*dto.CartsResponse, err error) {
+	cartedCourses, err := u.cartRepository.FindAll(userId)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, cc := range cartedCourses {
+		cart := &dto.CartsResponse{
+			CourseId: cc.CourseId,
+			Title: cc.Course.Title,
+			ImgThumbnail: cc.Course.ImgThumbnail,
+			AuthorName: cc.Course.AuthorName,
+			Price: cc.Price,
+		}
+		carts = append(carts, cart)
+	}
+	return carts, nil
 }
